+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/ui/MainActivity.kt
@@ -1,10 +1,10 @@
-package com.mani.quotify007.presentation
+package com.mani.quotify007.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.mani.quotify007.presentation.ui.theme.QuotifyAppTheme
-import com.mani.quotify007.presentation.view.MainScreen
+import com.mani.quotify007.ui.navigation.MainScreen
+import com.mani.quotify007.ui.theme.QuotifyAppTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
