+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/ui/navigation/model/MainState.kt
@@ -1,0 +1,9 @@
+package com.mani.quotify007.ui.navigation.model
+
+import com.mani.quotify007.domain.model.Quote
+
+data class MainState(
+    val quotes: List<Quote> = emptyList(),
+    val favoriteQuotes: List<Quote> = emptyList(),
+    val randomQuote: Quote? = null
+)
