+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/ui/screens/bottomappbar/BottomAppBar.kt
@@ -1,4 +1,4 @@
-package com.mani.quotify007.presentation.view
+package com.mani.quotify007.ui.screens.bottomappbar
 
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBarItem
@@ -6,7 +6,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavHostController
-import com.mani.quotify007.presentation.view.utils.BottomNavItem
 
 @Composable
 fun BottomAppBar(navController: NavHostController?) {
